Updated README example of API client
A naïve copy and paste of the example API client doesn't work, because `process_url` will create a URL without a slash between `github.com` and `users/myfreeweb`. 

Secondly, the GitHub API will refuse requests that do not have a User-Agent[1]. I see that this project has a `process_headers` override, but only for response headers. Being able to do this for request headers too would be helpful. I'm not knowledgeable enough yet to make this change cleanly. :)

[1]: http://developer.github.com/v3/#user-agent-required