Fix #100: Follow redirects when extended
API clients that are built by extending `HTTPotion.Base`
can use the `process_options` hook to ensure that all
redirects will be followed automatically. A typo in that
module caused `follow_redirects` to be ignored.

This adds a test to specifically cover the behavior of
an API client using `process_options` to inject a `true`
value for `follow_redirects`, and fixes the typo.